+++ tests/optimagic/parameters/test_scale_conversion.py
@@ -3,7 +3,7 @@ import pytest
 from optimagic import first_derivative
 from optimagic.parameters.conversion import InternalParams
 from optimagic.parameters.scale_conversion import get_scale_converter
-from optimagic.options import ScalingOptions
+from optimagic.parameters.scaling import ScalingOptions
 from numpy.testing import assert_array_almost_equal as aaae
 from numpy.testing import assert_array_equal as aae
